+++ app/assets/javascripts/app/views/single-post-viewer/single_post_content_view.js
@@ -1,6 +1,10 @@
 // @license magnet:?xt=urn:btih:0b31508aeb0634b347b8270c7bee4d411b5d4109&dn=agpl-3.0.txt AGPL-v3-or-Later
 
 app.views.SinglePostContent = app.views.Base.extend({
+  events: {
+    "click .near-from": "toggleMap"
+  },
+
   templateName: "single-post-viewer/single-post-content",
   tooltipSelector: "time, .post_scope",
 
@@ -10,8 +14,7 @@ app.views.SinglePostContent = app.views.Base.extend({
     "#real-post-content" : "postContentView",
     ".oembed" : "oEmbedView",
     ".opengraph" : "openGraphView",
-    ".status-message-location" : "postLocationStreamView",
-    ".poll": "pollView"
+    ".poll": "pollView",
   },
 
   initialize : function() {
@@ -23,8 +26,59 @@ app.views.SinglePostContent = app.views.Base.extend({
     this.pollView = new app.views.Poll({ model: this.model });
   },
 
-  postLocationStreamView : function(){
-    return new app.views.LocationStream({ model : this.model});
+  map : function(){
+    if (this.$el.find(".mapContainer")){
+
+      // find and set height of mapContainer to max size of the container
+      // which is necessary to have all necessary tiles prerendered
+      var mapContainer = this.$el.find(".mapContainer");
+      mapContainer.css("height", "200px");
+
+      // get location data and render map
+      var location = this.model.get("location");
+
+      // If map function is enabled the maptiles from the Heidelberg University are used by default.
+
+      var map = L.map(mapContainer[0]).setView([location.lat, location.lng], 14);
+
+      var tiles = L.tileLayer("http://korona.geog.uni-heidelberg.de/tiles/roads/x={x}&y={y}&z={z}", {
+        attribution: "Map data &copy; <a href='http://openstreetmap.org'>OpenStreetMap</a> contributors, " +
+                      "rendering <a href='http://giscience.uni-hd.de/'>" +
+                      "GIScience Research Group @ Heidelberg University</a>",
+        maxZoom: 18,
+      });
+
+      // If the mapbox option is enabled in the diaspora.yml, the mapbox tiles with the podmin's credentials are used.
+      if (gon.appConfig.map.mapbox.enabled) {
+
+        tiles = L.tileLayer("https://api.tiles.mapbox.com/v4/{id}/{z}/{x}/{y}.png?access_token={accessToken}", {
+          id: gon.appConfig.map.mapbox.id,
+          /* jshint camelcase: false */
+          accessToken: gon.appConfig.map.mapbox.access_token,
+          /* jshint camelcase: true */
+          attribution: "Map data &copy; <a href='http://openstreetmap.org'>OpenStreetMap</a> contributors, " +
+                       "<a href='http://creativecommons.org/licenses/by-sa/2.0/''>CC-BY-SA</a>, " +
+                       "Imagery © <a href='https://www.mapbox.com'>Mapbox</a>",
+          maxZoom: 18,
+        });
+      }
+
+      tiles.addTo(map);
+
+      // set mapContainer size to a smaller preview size
+      mapContainer.css("height", "75px");
+      map.invalidateSize();
+
+      // put marker on map
+      L.marker(location).addTo(map);
+      return map;
+    }
+  },
+
+  toggleMap: function () {
+    $(".mapContainer").height($(".small-map")[0] ? 200 : 50);
+    $(".leaflet-control-zoom").css("display", $(".small-map")[0] ? "block" : "none");
+    $(".mapContainer").toggleClass("small-map");
   },
 
   presenter : function() {
@@ -37,6 +91,10 @@ app.views.SinglePostContent = app.views.Base.extend({
 
   showPost : function() {
     return (app.currentUser.get("showNsfw")) || !this.model.get("nsfw");
+  },
+
+  postRenderTemplate : function(){
+    _.defer(_.bind(this.map, this));
   }
 });
 // @license-end