+++ app/assets/javascripts/main.js
@@ -45,3 +45,4 @@
 //= require osmlocator
 //= require bootstrap-switch
 //= require blueimp-gallery
+//= require leaflet