+++ app/controllers/application_controller.rb
@@ -145,7 +145,8 @@ class ApplicationController < ActionController::Base
   def gon_set_appconfig
     gon.push(appConfig: {
                chat:     {enabled: AppConfig.chat.enabled?},
-               settings: {podname: AppConfig.settings.pod_name}
+               settings: {podname: AppConfig.settings.pod_name},
+               map:      {mapbox: AppConfig.map.mapbox}
              })
   end